fix: move autocompleteQuery to coreService
